refactor : TestService Interface를 Class로 변경
